The big merge...
git-svn-id: https://unknownworlds.svn.cloudforge.com/ns1@21 67975925-1194-0748-b3d5-c16f83f1a3a1

+++ main/source/cl_dll/CLabelHeader.h
@@ -119,7 +119,7 @@ public:
 		char buf[512];
 		strcpy(buf, text);
 
-		int len = strlen(buf);
+		size_t len = strlen(buf);
 		while (len && isspace(buf[--len]))
 		{
 			buf[len] = 0;
